Split up and replace DefaultConfigSource
