+++ src/test/java/com/statemachinesystems/envy/DelegatingConfigSourceTest.java
@@ -1,6 +1,7 @@
 package com.statemachinesystems.envy;
 
 import com.statemachinesystems.envy.common.StubConfigSource;
+import com.statemachinesystems.envy.sources.DelegatingConfigSource;
 import org.junit.Test;
 
 import static org.junit.Assert.assertThat;
@@ -35,5 +36,23 @@ public class DelegatingConfigSourceTest {
         ConfigSource configSource = new DelegatingConfigSource(source1, source2);
 
         assertThat(configSource.getValue(new Parameter("baz")), nullValue());
+    }
+
+    @Test
+    public void looksUpValuesInNestedConfigSources() {
+        ConfigSource child1 = new StubConfigSource().add("foo", "1");
+        ConfigSource child2 = new StubConfigSource().add("bar", "2");
+        ConfigSource parent1 = new DelegatingConfigSource(child1, child2);
+
+        ConfigSource child3 = new StubConfigSource().add("foo", "3");
+        ConfigSource child4 = new StubConfigSource().add("bar", "4");
+        ConfigSource child5 = new StubConfigSource().add("baz", "5");
+        ConfigSource parent2 = new DelegatingConfigSource(child3, child4, child5);
+
+        ConfigSource configSource = new DelegatingConfigSource(parent1, parent2);
+
+        assertThat(configSource.getValue(new Parameter("foo")), equalTo("1"));
+        assertThat(configSource.getValue(new Parameter("bar")), equalTo("2"));
+        assertThat(configSource.getValue(new Parameter("baz")), equalTo("5"));
     }
 }
